+++ app/javascript/vue/tasks/uniquify/people/store/actions/findMatchPeople.js
@@ -6,7 +6,7 @@ export default ({ state, commit }, params) => {
   state.mergeList = []
 
   People.where({ ...params, extend: ['roles'] }).then(response => {
-    commit(MutationNames.SetMatchPeople, response.body)
+    commit(MutationNames.SetMatchPeople, response.body.filter(p => state.selectedPerson.id !== p.id))
     state.URLRequest = response.request.responseURL
   }).finally(_ => {
     state.isLoading = false